+++ Common/Requests/UpdatePrice.cs
@@ -1,4 +1,4 @@
 ﻿namespace Common.Requests
 {
-    public record UpdatePrice(long sellerId, long productId, decimal price, int instanceId);
+    public record UpdatePrice(int sellerId, int productId, float price, int instanceId);
 }
